update mark and users at the same time

+++ src/com/chf/action/DeleteUserAction.java
@@ -2,6 +2,7 @@ package com.chf.action;
 
 import com.chf.entity.User;
 import com.chf.service.UserService;
+import static java.lang.Integer.parseInt;
 
 import java.util.List;
 
@@ -9,17 +10,26 @@ public class DeleteUserAction extends Action{
 	private static final long serialVersionUID = 4878247652427455524L;
 	private UserService userService;
 	private String userName;
-
+    private String role;
 	@Override
 	public String execute()
 	{
         userName = request.getParameter("userName");
+        role = request.getParameter("role");
+
 		userService=new UserService();
 		User duser=new User();
-		duser.setUserName(userName);
-		this.userService.deleteUser(duser);
+
+        duser.setUserName(userName);
+
+        if(role!=null){
+            duser.setRole(parseInt(role));
+            this.userService.deleteUser(duser);
+        }
+
         List<User> users=userService.findUsers();
         request.setAttribute("users",users);
+
         return "updateUsers.jsp";
 	}
 }
